[diagnostics] Make filters match a prefix
Rust ends up generating tags of the form foo::bar, but the filter will
only do an exact match. This CL changes it so that it will do a prefix
match where tags are of the form foo::bar.

This addresses an issue for Fxfs where we can see top-level messages
sent to the console, but no messages below that.

Change-Id: Ib5c6e4246d9f21a69f5524cf46fdcae5d2647211
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/665504
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>