[input] Prepare fuchsia.iu.focus.FocusChainListenerRegistry for rerouting
The route for this FIDL protocol needs to go into the `/core/ui` realm
instead of into
`/core/session-manager/session:session/workstation-session`.  Since the
current route is partly in the separate repo `//src/experiences`, we
need to move the cut-over point to `fuchsia.git` to soft-migrate, since
that would place the cut-over change fully in `fuchsia.git`.

This change has a companion in `//src/experiences` which needs to go in
after. See:
https://fuchsia-review.googlesource.com/c/experiences/+/666329

Bug: 95901
Change-Id: I45dc57fd4a42f07faa8f3a3aa79234bafee3aa9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666850
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>