[component][fuzzing] Add framework fuzzer-test integration test
This CL adds a fuzzer-test as an integration test. Normally, the GN
templates for component fuzzing (added in a subsequent CL) will generate
both a fuzzer as well as a "fuzzer-test". This test does not use
coverage or mutate inputs. Instead, it verifies the fuzzer can build,
link, and run with a set of fixed inputs. This makes it appropriate for
CQ and helps mitigate bit rot.

This CL explicitly creates a similar "fuzzer-test", but here it is used
to verify the test-engine used with such fuzzer-tests works properly.

In fact, several aspects of building fuzzable target processes have not
been covered by any other tests, and bit-rotted as a result. As a
consequence, this CL also fixes those target library classes as well.

Bug: 92490
Change-Id: Ie7f2961ad1a9507792ad07bfd0bc31201cf1af1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662574
Reviewed-by: Nick Cano <nickcano@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>