[fshost] Don't set PATH for arbitrary programs
This line was added when devmgr was responsible for launching programs
in https://fxbug.dev/28172 but appears to have become detritus some time
ago.

Note that even before this change `which ls` reports that $PATH is unset
in both the console shell and via netruncmd. Making `which` work was
among the stated goals of the change that introduced this line.

It appears this was initially done in response to
https://crrev.com/c/521724, where Chromium tests that expected PATH to
be defined were updated to accommodate for its absence on Fuchsia. The
accommodation was removed after this addition in
https://crrev.com/c/766470. 8 months later, the accommodation was added
back in https://crrev.com/c/1141929 without explanation (probably
because devmgr was no longer launching whatever is running those
Chromium tests).

Remove unused dependencies on //zircon/system/ulib/zircon-internal.

Change-Id: I92dffeba2c3f069e8312e66ff26370a8a57fee6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666747
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>