[kernel] Strip tags in FAR before passing to kernel accesse fault handler
As a followup to fxr/659469, we should also strip the tag when handling
access fault exceptions before they reach the VM's access fault handler.
From what I could tell, there may not be an easy way of toggling the
accessed flag from userspace, but we can manually remove the flag when
aging pages. The `vmaspace_accessed_test` test does this. This CL
reuses some of the machinery for the test such that we can load from
tagged user pointers on arm. The ensures that the FAR is tagged on
access faults.

Bug: 81499
Change-Id: I11cb76dae83b057bc1360296fabb47969629d93a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663583
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>