[hypervisor] Add IdAllocator::Reset
This allows us to dynamically set the size of an IdAllocator to a value
less than or equal to the statically defined maximum.

This is useful in the case where we statically know the maximum possible
value allowed, but must dynamically query the hardware for what limits
it supports.

Bug: 95763
Change-Id: I0a7a972aef8dcdf2d7fdf7452114d102a8f8d5a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666435
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>