[starnix][binder] Restructure locking and waiting
Moves several related fields of data in BinderThread under a single
lock, making it easier to reason about race conditions.

In addition, the waiting mechanism is improved for threads. A WaitQueue
is not necessary, as only a single waiter is ever waiting on a
BinderThread's command queue.

The WaitQueue is needed at the BinderProcess level, so that threads can
wait for events on the /dev/binder FD.

Change-Id: I537a734bceba503c84b193568d41a2972430af45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666965
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>