[pwauth] Drop CredentialManager-managed credentials on account removal
Expand the KeyEnrollment trait to handle both key provisioning and
removal, and then implement removal as no-ops for null auth and
scrypt, and as a call to Credential Manager for pinweaver.

In the event of a failure from Credential Manager, we ignore the failure
and carry on with the account deletion, since our threat model is still
secure against even a malicious Credential Manager provided the user's
password remains secret.

Test: in addition to unit tests, I also performed the following manual
end-to-end test:
* applied the following patch in the series to enable pinweaver and OTAd
* created a new account
* verified via the shell that credential_manager's data directory
  contained a new subfolder under creds/ with three files
* removed the account via `accountctl remove-all`
* observed that credential_manager's data directory no longer contained
  the subfolder of creds/ that had been present before

Fixed: 89078

Change-Id: I04fa8d60c32b9ef9eef4b8362af20a260af14693
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666561
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>