[fidl][cpp] Strict unions cannot be default constructed
By making this change, we avoid having to answer questions such as "what
is the default state of a strict union" and "what happens when the user
calls Which on that union". Now these cases are ruled out at compile
time.

Of course, within the FIDL we still need to make those unions during
decoding. That ability is exposed through the
|fidl::internal::DefaultConstructPossiblyInvalidObject| trait, which end
users won't be able to use.

The main hurdle was adjusting the various codes that materialize FIDL
objects to use this trait, and making this trait work for arrays of
(arrays of (...)) strict unions.

It adds some subtle C++ complexity but makes the end user API simpler.

Change-Id: Ic74e5e351891d65d8c599614ccc4f041517dd86e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666560
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>