[starnix] Refactor task interruption and fix exit code race
Task interruption was previously done by abusing the SIGKILl signal.

Instead introduce Task::interrupt that will interrupt a running task and
loop back to the exception loop.
To be noted: This currently only have effect when the task is blocked on
a syscall as the current implementation does not interrupt user-space running code. This
will be changed when it is moved to the new execution model.

This CL also fixes the race between exit and exit_group by ensuring exit
does not override the exit status of a task.

Change-Id: I0187af39a7b80231c906c4c8dbeeebcfd4335425
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667522
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>