[sl4f][cfv2] Delete rotated_square_test.
This test is not being run on any builders despite being included in
src/ui/scenic:test. This is because the rotate_square_test is written as
a host-driven test using SL4F, and it currently builds only as a
dart_library. So I think the builders just builds the test, but won't
run it.

This will allow us to remove the sl4f.scenic.PresentView facade.

Bug: 77056
Change-Id: I12c15418d308ec4b95c54952d66075fd48deec89
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667684
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Yuan Zhi <yuanzhi@google.com>