[products] Introduce hardware_testing product config
This CL removes the hardware_testing bundle and replaces
it with the hardware_testing product config.

This config is presently based on core with several
additional utilities included for the purposes of hardware
testing.

The next step is to set up automated builders. For now
I tested this by building at my workstation, flashing
onto a device and verifying that the necessary utilities
exist in the package cache.

Change-Id: Id3e86fb4ddf7217c740a312adb5474e9b5105436
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667631
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>