[rust] Generate a depfile for the toolchain copier
We can't rely on the .cipd_version file existing in all cases where we
want to copy the toolchain, so we have to generate a depfile detailing
all of the files we copy.

This also fixes the hermeticity checker by adding dependency edges from
the toolchain source files to the toolchain output files.

Change-Id: Id02bff26894b7e363dec2b9cbef56ec1f6b230d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666851
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: David Koloski <dkoloski@google.com>