[insntrace][flake] Ensure device closed before next open
A flake manifested due to the fact that close was not performed
synchronously. The insntrace device only allows for a single client
connection at a time to be open and it was possible for it to observe to
open calls before seeing the close. This race has always been inherently
present, but a recent change must have introduced timing changes which
made it more likely to occur.

MULTIPLY: insntrace_tests
Fixed: 97097
Change-Id: If6b4ea5c04475fbba88db10d214eebd3e0f7f59e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666955
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>