[zxdump] Skeleton dump API & gcore command
This implements the skeleton of the dumper logic and the
command-line utility using it.  It just writes empty ELF headers.

The test harness code is extended to sandbox a Fuchsia tool
process so it gets only a private /tmp mapped into a subdirectory
of the test's own /tmp and a private /svc that fakes the
fuchsia.kernel.RootJob protocol with a different job handle.

Change-Id: Ic69a54559f94ec2bba14a901c0d3a86708b3a403
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542803
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>