[serve-updates] Add warning when default device is not set
When the default device is not set via fx set-device, fx
serve/serve-updates get stuck on "Discovery..." loop.

There is no actionable error message to the user on how to resolve this
issue. Many users resort to chat or other support channels which results
in lost developer productivity.

The better fix would likely be to clean up the logic in serve-updates
but that may require most testing and investigation to ensure it does
not affect existing developers and integrations.

Note: this also cleans up the message from ffx repository when ffx is
used as repository serving mode and triggered by fx serve/serve-updates.

Bug: 97694
Test: fx serve with fx set-device empty and set.
Change-Id: I8ca5deb335dc866ac469f4b1655a11f62337c2a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668162
Reviewed-by: Gary Boone <gboone@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Amit Uttamchandani <amituttam@google.com>