[scrutiny] Remove logo from the shell.
This change removes the old logo from the shell. This is a very old
bit of code that was used in some of the initial prototypes. As the
shell becomes involved in more automated processes the logo just
adds extra noise to logs we don't need.

Bug: 83871
Change-Id: Icdbd648395cfe6ddfba1296b71443bf5d8fee260
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668443
Fuchsia-Auto-Submit: Ben Wright <benwright@google.com>
Reviewed-by: Daniel Hwang <arkay@google.com>
Commit-Queue: Ben Wright <benwright@google.com>