[test_manager] Save inspect when a run controller is closed
Takes a snapshot of inspect for a test run when the run controller
is closed. When a run controller channel is closed, this is often
because running some tests has timed out. Saving inspect in these
situations will allow us to debug these cases.

Note that this change also slightly alters teardown in the case a
user terminates ffx test or run-test-suite. Test manager will now
immediately detect the channel is closed instead of needing to poll
for the next request off the channel to start teardown.

Bug: 95923, 94333
Fixed: 97012
Testing: New unit tests. Manually verify that terminating ffx test
while a test is running results in the inspect for that test being
persisted. Manually verify that allowing ffx test to run to completion
does not result in inspect being persisted.

Change-Id: Ia427193f46e2849b7ca80c7622fbbe07a8806e98
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667442
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>