[CredMgr] Remove or localize allow(dead_code).
Note a chunk of the dead code was for checking the consistency of the
hash tree. I added a consistency check on load to use the functionality
although don't know of any conditions that could cause it to fail.

Also fixed a couple of very small bugs while here.

Change-Id: Icdd07d5a48817b6300ed474cfb8b120d8bb335e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667912
Reviewed-by: Ben Wright <benwright@google.com>
Reviewed-by: Daniel Hwang <arkay@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>