[gigaboot] Initialize UART ZBI entry
Construct a UART ZBI entry from the SPCR ACPI table and appends it to
the ZBI. This requires adding the memory range containing the UART MMIO
base address, as physboot expects the memory range to contain all MMIO
addresses.

Also disables the caches and MMU prior to jumping into Zircon, as the
kernel asserts that these are not initialized.

Bug: 97287
Change-Id: Ic123782a457e71abb5661f9ed4755c0241b5350a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666962
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>