[boot-shim] Add UartItem
Unlike AcpiUartItem the purpose of this item is to hold and translate
any uart config. This allows for ACPI parsing to happen outside of the
item construction, such that the serial configuration can be extracted
earlier during the boot process and enabling stdout (e.g.
LegacyBootShim).

Test: fx test boot-shim-tests
Change-Id: I3877482e1b4f33dfbe2dcf0321bb913300ddaea0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666843
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>