[tiles][flatland] quality of life improvements
Put a background behind the tiles and maintain a margin between
them.

Also notice when an app has exited and remove its tile, and, when
removing a tile, stop the app from running.

In order to keep the exiting unit tests, added an optional
value for the padding and don't use it in the tests.

Bug: 80814
Bug: 88623

Testing: manual plus unit tests.

Change-Id: I40dd8b4cbd18b23e93289a8a55a30ea88ac6ffc5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668124
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Rob Tsuk <robtsuk@google.com>