[nand][drivers] Resurrect AS370 nand driver.
Brings back the previously deleted driver as we consider returning
support for it. This is being brought back in as minimal a manner as
possible, updated deps and paths to match migrations that have happened
and left original copyright dates intact. Incremental work on the driver
to follow.

Migrates the one lifetime test to MockDevice as fake_ddk is now
restricted use.

Bug: 95994
Change-Id: Ie0299d95416ac3fc0b159c18c4d4d8f6552e9d88
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660642
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>