[CredMgr] Introduce diagnostics trait.
This CL introduces the most basic form of a new diagnostics trait to
record diagnostics events. The trait currently only contains one method
to record the outcome of FIDL methods (i.e. method name plus either
success or the type of error).

Their are two implementations of the trait: a fake that is used in unit
tests (doing this required an additional test to get coverage of the
`handle_requests_for_stream` and `handle_request` methods) and an
inspect implementation that is used in production but is currently
non-functional until the next CL!

Bug: 91714
Change-Id: I43bccb74cce35613cfddbe267c1dbb8715b5a63a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668416
Reviewed-by: Ben Wright <benwright@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>