[driver_runtime] Add support for delayed tasks
This patch implements support for delayed for tasks by multiplexing a
single delayed task which we post to the shared queue as if it were a
timer object. Unlike a timer, it unforuntately lacks that ability to be
updated, which will cause a slight performance overhead if multiple
concurrent delayed tasks are scheduled on a single dispatcher where the
tasks scheduled later have a sooner deadline than the first. This is in
all likelihood going to be rare as we delayed tasks are not widely used.

I opted to avoid forwarding each outstanding wait with a wrapper to the
shared dispatcher to simplify complexity around cancelation and
shutdown and avoid extraneous allocations.

MULTIPLY: driver_runtime_tests
Test: fx test driver_runtime_tests
Change-Id: I466d872ab30100392a0a003f5f0c1cba0680e896
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668586
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>