[phys] Trampoline boot relocates data ZBI
When a fixed data load address is passed, the data ZBI is relocated
as well, and the zbi pointer passed to the next kernel item is updated.

This CL is the continuation of the plumbing. Here the Trampoline
code is extended to conditionally copy the data ZBI, when a load
address is provided.

The trampoline test is updated to verify that the boot zbi address
matches the fixed data address.

Also TrampolineArgs is allocated as part of the Trampoline. Since
when doing multiple coies(> 1, due to data and kernel being relocated),
we cannot guarantee that the shim's stack wont be clobbered.

Bug: 88583
Test: fx run-zbi-test trampoline-boot-zbi-test
Change-Id: Iafaea02b7195de85e706a5b87f10d8845f6fe50f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652245
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>