[fidl] Transport-dependent iovec array size
Currently, the iovec buffer in UnownedEncodedMessage has a fixed size
regardless of what transport is being used.

This CL adjusts the size based on transport.

This is based on a bug that was discovered with the driver transport
after the FIDL encoder was changed to an alternative encoder.

Change-Id: Ide20c79a5a3d16ddd6aad633989384ef2e3222c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668404
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>