[driver_runtime] Split Idle event into CompleteShutdown & Idle.
I've changed fdf_internal_wait_until_idle to not wait for waits
and future scheduled tasks.

This is as in the driver host v2 tests, we start up a driver and
want to wait until idle, but it will never succeed if it has called
begin wait but not expected it to be signaled yet.

There are now 2 separate notifications - one for when the dispatcher
enters an idle state (waits_ + timer_ not checked), and the other when
the dispatcher is ready to shutdown (idle state + waits_ empty +
!timer_.is_armed).

Bug: 87162

Change-Id: Ie6191bac6b1bc11d10c4966bb8eb1b9ca0a6b831
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668856
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>