[tools][test-list] Add tags from test-list.json to summary.json.
In order to have the test metadata from test-list.json in resultdb, we
can copy it through to the summary.json which we process to upload test
results to resultdb. We can then add those tags to the list of tags we
attach to each test result that we upload.

Bug: b/228092454
Change-Id: I3de9a5bf620d8616eabed0d79c5259f614c480e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668419
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>