[fdf][examples] Write DFv2 demo-number example
This is going to be the example driver for DFv2.
Start by building the driver with the current DFv2
libraries, but this example will be updated to
use best practices and SDK only libraries.

This change was originally reverted because it broke
the build, but the build errors were fixed.

Bug: 97855
Change-Id: I1125afb69ad6a1d2db353f4c6186fe1d6a860244
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669042
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>