[phys] Introduce phys/acpi.h
This provides a mechanism for obtaining a AcpiParser meant
to be used on phys environment, for handling translation from
virtual to physical addresses.

PhysMemReader has no state, so we choose to provide a static reader
that is used by all instances of the parser.

No functional changes.

Bug: 97896
Test: builds and CQ
Change-Id: I6f90038d1ef03f113e904a14bc1e6eccd62f5994
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666849
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>