[gigaboot] Construct GIC configuration from ACPI
Uses the MADT table to construct either a GICv2 or GICv3 configuration.
We could likely configure a large portion of the x64 interrupt
controller with the same approach, but once again that's autodetected
in the kernel so we don't need to do that here.

Bug: 97287
Change-Id: Ie993c0edfe2901f1ed7b856674a4ac13de01e3eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667512
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: David Pursell <dpursell@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>