[weave] Enable weave-ip-fwd-test-wpan-subnet-6
This change updates the prefix length for addresses assigned to the
weave interfaces so that the wlan and wpan nodes are not considered
directly connected to the weave-node. Without this change, weave-node
will not attempt to route packets to the wlan or wpan node through
the fuchsia-node (gateway).

Change-Id: Ia654c84ad8db3daec8409ca7108dd439bab4d448
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667244
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>