[events] Use a singleton Hooks object.
The hook system used a Hooks object per component instance
with a reference to the parent.

When an event was dispatched, the event would propagate through
each parent. This system is unnecessary now.

The only hooks installed are the ones at the root using
builtin environment.

Now, we can simply create a singleton Hooks object and share it
between all the component instances.

Test: fx test component-manager-tests
Change-Id: I133c6c4f332116321ec02dddded70db3f878f08e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668094
Reviewed-by: Brian Bosak <bbosak@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>