[tefmocheck] Respect except blocks even if split across tests.
When a failure check has AttributeToTest: true, sometimes the start
string of an except block can show up in one test's output and the end
string shows up in the next test's output. Since we only check the
output for one test at a time, we don't see the end string and so we
count the failure string as a failure when it should be ignored.
Instead, we'll check the entire output log and only consider the failure
strings within the block per test but still check whether it's contained
within an except block in the whole log.

Bug: 97776
Change-Id: Ie10c889e65a4c50b9e9aa84fc7a3b1394c34bbe8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668746
Reviewed-by: Leonard Chan <leonardchan@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>