[network/tests] Remove dhcp_interop test
This test requires netemul guest support to be migrated to CFv2, which
has yet to be built out. Given this is the only test that uses the
netemul guest functionality, opt to remove it, with the option to revive
it as a v2 test once the support exists.

Fixed: 97735, 96132
Bug: 87587
Change-Id: I050ef6cf3ff825bf0de4431f677b9f6e5e192070
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667707
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>