[basemgr] Rate limit device reboots
This CL modifies basemgr to rate limit reboots triggered
by crashes of the session shell. Prior to this CL, basemgr
would trigger a system reboot if the session shell crashes
4 times within a 1 hour window. This behavior is largely kept
the same except that now the reboot attempts themselves are
rate limited with an exponential backoff. The growth rate
of this limit is by powers of 2, e.g. 2, 4, 8, measured in
minutes since last reboot. The backoff is capped at 64 minutes.
Furthermore, after one day, the backoff resets to 0. Lastly,
logic is added to ensure that this only executes if and when
the system clock has started.

Bug: 82401

Change-Id: I32a90f7eb4770639450967b7b09e36f22345fa2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667802
Reviewed-by: Rich Kadel <richkadel@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>