[ffx] Make unknown names unique
We sometimes show multiple targets with the name <unknown>. This adds
some code so that in those cases we show something like <unknown-1> and
<unknown-2> instead. The actually index that we use is the index into
the list of total targets so if it is the first target in the list it
would get <unknown-0> for example.

There is a higher level question about how to make this all better but
this is a good enough fix to help alleviate some pain.

Fixed: 93823

Change-Id: Icdace48441a84ff7182df827fd0aadd1f2105037
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667002
Reviewed-by: Omer Lev-ran <omerlevran@google.com>
Commit-Queue: Amit Uttamchandani <amituttam@google.com>