[ui][input-pipeline] rename PointerMotionScaleHandler
We're soon going to add another handler which scale pointer
motion, so the name of this handler is going to become
ambiguous.

Accordingly, we rename PointerMotionScaleHandler to
PointerMotionDisplayScaleHandler. This name specifies
the concern that we scale for.

This is purely a rename. There are no functional
changes.

Bug: 88200
Test: fx test input_pipeline_lib_tests
Change-Id: I9f3968cb9507ce73e384071c192eb9dd1823cc3f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669676
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>
Commit-Queue: Mukesh Agrawal <quiche@google.com>