[ui][input-pipeline] add immersive-mode shortcut handler
Refactor the immersive mode shortcut detection, placing that
logic in a standalone handler.

This motivated by the fact that we will the pointer
acceleration logic to be aware of immersive mode as well.

More specifically: pointer acceleration should be disabled
when in immersive mode, since pointer acceleration makes
some games more difficult to play.

Note that this is _not_ a pure refactor. The old code
is was more permissive of what was accepted for the shortcut,
recognizing the shortcut when additional modifiers were
actuated.

For example: the old code would accept CTRL-ALT-SHIFT-i,
or CapsLock-Alt-Shift-i. The new code does not.

Along the way:
- Update the Scene Manager to install the new handler.
- Update the MouseInjectorHandler to use the event emitted
  by the ImmersiveModeShortcutHandler, instead of detecting
  the shortcut on its own.
- Derive `Copy` for `input_device::Handled`, since that
  enum is lightweight, and we expect it to stay that way.
  (This avoids the need for an explicit clone() call.)
- Add KeyboardEvent::get_unsided_modifiers(), for callers
  who are interested in the semantics of the modifiers,
  but not their positions.

Bug: 88200
Test: fx test input_pipeline_lib_tests
Test: manual (see below)

Manual test
-----------
OTA'd an atlas with these changes, verified that alt-shift-i
hid the mouse.

Change-Id: I741fa1322eba124ae322fe3692fb6099e37a48d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669677
Reviewed-by: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>