[fidlgen][dart] Support unknown interactions.
Update the Dart bindings to generate support for both sending and
handling unknown interactions on both the server and client side.

These changes to the bindings are only active when fidlc supplies
the new openness/strictness modifiers.

This is a prototype, so the design choices for how the bindings handle
supporting unknown events and unknown method calls are not finalized.
This initial setup demonstrates how the changes could work and provides
a baseline of tests that cover the changes.

Test: fx test fuchsia-pkg://fuchsia.com/dart-bindings-test#meta/dart-bindings-test.cm
Bug: 88366
Change-Id: Ic677a219202fb9d765ec81cf7e16c8aaaf5db785
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663743
Reviewed-by: Ian McKellar <ianloic@google.com>
API-Review: Ian McKellar <ianloic@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>