[phys] trampoline-boot-test random seed
ZBI_TYPE_SECURE_ENTROPY is now always appended for the test.
This provides a fallback source for randomness. First we will try
using arch::Random if supported. If this is not the case, then
we will default to using the first 8 bytes of the first secure
entropy item available whose size is at least 8(sizeof(uint64_t)
and if that is not available, we will look if any boot option
is available from the command line as a last resot.

This only happens in the bootstrap step, since any other seed
value is propagated forward as it is updated through different
iterations.

The seed is always logged, for reproduction use:
trampoline.user_seed=$random_seed

Failure to find at least one source of randomness will trigger
an assertion.

Bug: 88583
Test: fx run-zbi-test trampoline-boot-test
Change-Id: I2fa005f8a8326b80c283f28d6947889353876029
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655188
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>