[run-test-suite] Test that streamed artifacts are polled
Adds an integration test that verifies that streamed artifacts
are polled at the same time tests are run. This is important as
artifacts such as logs need to be shown to the developer as they
arrive, rather than once at the end.
To test this, we add a writer that sends asynchronous signals when
certain strings are written. Then we run a hung test, and verify that
any output it produces is visible before cancelling the test.

Testing: new unit and integration test
Multiply: run_test_suite_integration_tests

Change-Id: Ib60959d28ad8e51556dd8dfc4bdd5cbe1c1c8331
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669771
Reviewed-by: Ankur Mittal <anmittal@google.com>
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>