[perftest] Change tracing initialisation to be non-racy
Change the perftest library's trace provider startup code to avoid
race conditions in which early trace events can be lost.  That is done
by changing StartTraceProvider() to wait for tracing setup to complete
before returning.

The logic for doing that is somewhat complicated, so in the future it
would be good to move StartTraceProvider() to make it easily usable by
other code.

This logic is loosely based on CreateProviderSynchronouslyAndWait()
and WaitForTracingToStart() in
garnet/bin/trace/tests/integration_test_utils.cc.

Add a test for this which checks for trace events produced by the
perftest library.

Bug: 22911
Change-Id: If35a97edf68af6c3c7fbb00c5fdd50c7236fac3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/364515
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>