[starnix] Add the ability to run host tests
This CL makes it possible to write tests for starnix in C++ using gtest
in the Fuchsia source tree. These tests are built against the Fuchsia
Linux host toolchain and then run in starmium.

This CL adds a syscalls_test binary, which we can add gtest cases to as
we work on Starnix. The CL also shows how to run the existing fdio-test
host test using this mechanism. For now, the CL enables only a few tests
for this binary. We can enable more of these tests in a later CL.

Test: syscalls_test, fdio-test
Change-Id: I3871e5ccbf52d7401ec1d9a00627f01ae26f98a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669947
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Adam Barth <abarth@google.com>