[ffx] Add timestamp to manual targets
The attached bug is one of a few where the interaction between the FFX
target collection and the FEMU emulator lead to confusion. An emulator
running with SLiRP networking is undiscoverable by the FFX daemon, so a
manual entry must be added. But manual entries persist indefinitely,
leading to old entries piling up in the `ffx target list` output unless
they're manually removed. The `ffx emu` plugin has code to do that
manual cleanup, but terminating the emulator via `ffx emu` is not
required.

This CL implements one of the designs discussed in
go/ffx-target-collection-cleanup, which adds a timeout to the existing
manual target structures. Specifying this timeout is optional, and
indeed the `ffx target add` plugin leaves it off, so the behavior of
targets added from the command line is unchanged. For emulators, we add
a timeout of 120 seconds. The purpose of this timeout is to allow the
target collection sufficient time for the emulator to finish booting and
establish an RCS connection. Once established, the target behaves as
before.

When the target disconnects (i.e. the emulator terminates), the manual
target with a timeout is now allowed to expire as other RCS entries
would; this means emulator targets are automatically purged from the
target collection after termination. If the emulator never establishes
the RCS connection, the target will be removed after the timeout has
elapsed, so bad targets won't clutter the collection and `list` output.

In case of daemon restarts, the timeouts are included in the persistent
storage of manual targets in the FFX config. These are compared with the
present time whenever the daemon reloads manual targets, and expired
targets are excluded and removed.

Bug: fxbug.dev/94214

Change-Id: I59e0a0b0da3b4d941f5b3c5d081d6adee2e496da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667165
API-Review: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>