[wlan] Begin to use security feature query in SME.
This change plumbs the security support data into SME, and uses it for
SME's internal logic for determining if WPA3 is supported. A future
change will replace the rest of the uses of the old driver features
enum with the security support query in SME. This test passes the ACTS
test suites PolicyScanTest and VapeInteropTest.

Bug: 88315
Test: fx test wlanstack-tests wlan-hw-sim-test wlan-sme-tests
Change-Id: I869eba3b47b5ca5ef3d20bc78dcb0a457fb7f722
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669849
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Commit-Queue: Karl Ward <karlward@google.com>