[fdf][ephemeral] Add DFv2 registrar to driver mgr
Previous change added the DriverRegistrar capbility to the driver index.
We still want this to be consumed through the driver manager so we don't
expose the driver index. Add a registrar service to the driver manager
to forward requests to the driver index's registrar server. Also allow
the LoadDriver function to use a universe package resolver in order to
resolve these ephemeral drivers.

Bug: 95587
Tests: Unit test added
Change-Id: I7e223cf79d374ccd20d551624b89ce307b9263c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669215
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>