[kernel][vm] Minorly reorganize AllocateCopyPage() for clarity
The control flow in this function was not structured in a particularly
clear way, and could give off the impression that a variable was
mistakenly used in an uninitialized state. While this was not true, it
serves to restructure the logic a bit for clarity.

Fixed: 97879
Change-Id: I68a84e4e4d07d185e27845abdf584cec9a5387b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669962
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>