[CredMgr] Refactor inspect topology
This does a few things to make the structure a bit more scalable:

* The IncomingMethod enum is now constructed with a pro macro that lets
  inspect iterate over its values so we can pre-polutate nodes for each
  RPC and also retain more of the convention of snake_case inspect
  (thank you paste!{})
* We nest the actual error counts and include a summary count of the
  errors which makes it easier to write a traige rule to detect errors
* All the incoming RPC are nested under an "incoming" node to provide
  room for other types of output.

Looking at the inspect unittests is the easiest way to visualize the new
output.

Note being able to refactor inspect output like this without having to
change operational code or tests shows the benefit of isolating with the
diagnostics trait.

Bug: 91714
Change-Id: I1c6a3c752578edd04cc97eb5d21950089d7fd2a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669166
Reviewed-by: Ben Wright <benwright@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>