Revert "[dwc3] Delete unused driver."
This reverts commit bc20f2681a42701797135f417396da0a5aaa1537.

Un-delete an old driver for the DesignWare USB3 core.  Note that this
is currently dead code; it is not hooked into the build system, and
would not build even if you tried.

The next CL will restore the driver to a functioning state, as well as
adding some tests.  The primary purpose of this CL is to provide a
reference as to what changes were made to get the driver working again
after restoring.

Bug: 96652

Change-Id: Ie05b731060d9c60a5b65c0984b4a9ae5aef6d7ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667811
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: John Grossman <johngro@google.com>