[fx][serve] Fix erroring out with invalid port
In a prior refactor, we accidentally broke `fx serve-updates -l :8084`
from erroring out with the ffx repository server, and the server was not
configured to use that port with `ffx config set
repository.server.listen "[::]:8084"`. This patch changes the helper
function to not unilaterally exit early if the port happens to be free,
but also do the rest of the validation steps.

Fixed: 98017
Change-Id: Ie24837652ae48645d94550ff69e4b46148200cf3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670249
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>